+++ src/index.ts
@@ -119,6 +119,7 @@ export { useElementAspectRatio } from './hooks/useElementAspectRatio';
 export { useMeetingManager } from './providers/MeetingProvider';
 export { useVoiceFocus } from './providers/VoiceFocusProvider';
 export { useBackgroundBlur } from './providers/BackgroundBlurProvider';
+export { useBackgroundReplacement } from './providers/BackgroundReplacementProvider';
 export { useAudioVideo } from './providers/AudioVideoProvider';
 export { useRosterState } from './providers/RosterProvider';
 export { useRemoteVideoTileState } from './providers/RemoteVideoTileProvider';
@@ -164,6 +165,7 @@ export { RemoteVideoTileProvider } from './providers/RemoteVideoTileProvider';
 export { FeaturedVideoTileProvider } from './providers/FeaturedVideoTileProvider';
 export { VoiceFocusProvider } from './providers/VoiceFocusProvider';
 export { BackgroundBlurProvider } from './providers/BackgroundBlurProvider';
+export { BackgroundReplacementProvider } from './providers/BackgroundReplacementProvider';
 
 export {
   UserActivityProvider,