+++ src/providers/BackgroundBlurProvider/index.tsx
@@ -3,6 +3,7 @@
 
 import {
   BackgroundBlurOptions,
+  BackgroundBlurProcessor,
   BackgroundBlurVideoFrameProcessor,
   BackgroundFilterSpec,
   ConsoleLogger,
@@ -19,13 +20,16 @@ import React, {
   useEffect,
   useState,
 } from 'react';
+import useMemoCompare from '../../utils/use-memo-compare';
+import { isPrevNextUndefined, isPrevNextEmpty } from '../../utils/device-utils';
+import { BaseSdkProps } from '../../components/sdk/Base';
 
-interface Props {
+interface Props extends BaseSdkProps {
   /** The spec defines the assets that will be used for adding background blur to a frame. For more information, refer to
-   * [Amazon Chime JS SDK Background Filter Guide](https://github.com/aws/amazon-chime-sdk-js/blob/master/guides/15_Background_Filter_Video_Processor.md#adding-background-blur-to-your-application). */
+   * [Amazon Chime SDK for JavaScript Background Filter Guide](https://github.com/aws/amazon-chime-sdk-js/blob/main/guides/15_Background_Filter_Video_Processor.md#adding-a-background-filter-to-your-application). */
   spec?: BackgroundFilterSpec;
   /** A set of options that can be supplied when creating a background blur video frame processor. For more information, refer to
-   * [Amazon Chime JS SDK Background Filter Guide](https://github.com/aws/amazon-chime-sdk-js/blob/master/guides/15_Background_Filter_Video_Processor.md#adding-background-blur-to-your-application). */
+   * [Amazon Chime SDK for JavaScript Background Filter Guide](https://github.com/aws/amazon-chime-sdk-js/blob/main/guides/15_Background_Filter_Video_Processor.md#adding-a-background-filter-to-your-application). */
   options?: BackgroundBlurOptions;
 }
 
@@ -36,56 +40,78 @@ interface BackgroundBlurProviderState {
   isBackgroundBlurSupported: boolean | undefined;
 }
 
-const BackgroundBlurProviderContext = createContext<
-  BackgroundBlurProviderState | undefined
->(undefined);
+const BackgroundBlurProviderContext =
+  createContext<BackgroundBlurProviderState | undefined>(undefined);
 
 const BackgroundBlurProvider: FC<Props> = ({ spec, options, children }) => {
-  const [isBackgroundBlurSupported, setIsBackgroundBlurSupported] = useState<
-    boolean | undefined
-  >(undefined);
+  const [isBackgroundBlurSupported, setIsBackgroundBlurSupported] = useState<boolean | undefined>(undefined);
   const [processor, setProcessor] = useState<VideoFrameProcessor | undefined>();
+  
+  const blurSpec = useMemoCompare(
+    spec,
+    (
+      prev: BackgroundFilterSpec | undefined,
+      next: BackgroundFilterSpec | undefined
+    ): boolean => {
+      if (Object.is(prev, next)) {
+        return true;
+      }
+      return false;
+    }
+  );
+
+  const blurOptions = useMemoCompare(
+    options,
+    (
+      prev: BackgroundBlurOptions | undefined,
+      next: BackgroundBlurOptions | undefined
+    ): boolean => {
+      if (isPrevNextUndefined(prev, next) || isPrevNextEmpty(prev, next) || (prev?.filterCPUUtilization === next?.filterCPUUtilization)) {
+        return true;
+      }
+      return false;
+    }
+  );
 
   useEffect(() => {
-    async function initializeBackgroundBlur() {
-      try {
-        console.log('Initializing background blur processor.');
-        const createdProcessor = await BackgroundBlurVideoFrameProcessor.create(
-          spec,
-          options
-        );
-        // BackgroundBlurVideoFrameProcessor.create will return a NoOpVideoFrameProcessor
-        // in the case that BackgroundBlurVideoFrameProcessor.isSupported() returns false.
-        // BackgroundBlurVideoFrameProcessor.create() can also throw an error in case loading
-        // the assets are not fetched successfully.
-        if (createdProcessor instanceof NoOpVideoFrameProcessor) {
-          console.warn(`Initialized NoOpVideoFrameProcessor.`);
-          setProcessor(undefined);
-          setIsBackgroundBlurSupported(false);
-        } else {
-          console.log(
-            `Initialized background blur processor: ${JSON.stringify(
-              createdProcessor
-            )}`
-          );
-          setProcessor(createdProcessor);
-          setIsBackgroundBlurSupported(true);
-        }
-      } catch (error) {
-        console.error(
-          `Error creating a background blur video frame processor device.`,
-          error
-        );
-        setProcessor(undefined);
-        setIsBackgroundBlurSupported(false);
-      }
-    }
+    console.log('Specs or options were changed. Re-initializing background blur processor.');
     initializeBackgroundBlur();
     return () => {
       console.log(`Destroying background blur processor.`);
       processor?.destroy();
     };
-  }, []);
+  }, [blurOptions, blurSpec]);
+
+  async function initializeBackgroundBlur(): Promise<BackgroundBlurProcessor | undefined>{
+    console.log('Initializing background blur processor with ', spec, options);
+
+    try {
+      const createdProcessor = await BackgroundBlurVideoFrameProcessor.create(
+        spec,
+        options
+      );
+      // BackgroundBlurVideoFrameProcessor.create will return a NoOpVideoFrameProcessor
+      // in the case that BackgroundBlurVideoFrameProcessor.isSupported() returns false.
+      // BackgroundBlurVideoFrameProcessor.create() can also throw an error in case loading
+      // the assets are not fetched successfully.
+      if (createdProcessor instanceof NoOpVideoFrameProcessor) {
+        console.warn(`Initialized NoOpVideoFrameProcessor.`);
+        setProcessor(undefined);
+        setIsBackgroundBlurSupported(false);
+        return undefined;
+      } else {
+        console.log(`Initialized background blur processor: ${JSON.stringify(createdProcessor)}`);
+        setProcessor(createdProcessor);
+        setIsBackgroundBlurSupported(true);
+        return createdProcessor;
+      }
+    } catch (error) {
+      console.error(`Error creating a background blur video frame processor device`, error);
+      setProcessor(undefined);
+      setIsBackgroundBlurSupported(false);
+      return undefined;
+    }
+  }
 
   const createBackgroundBlurDevice = async (
     selectedDevice: Device
@@ -95,32 +121,20 @@ const BackgroundBlurProvider: FC<Props> = ({ spec, options, children }) => {
         selectedDevice
       )}`
     );
-    // It takes time for the processor to start - if you call this method before it is finished initializing, throw an error
-    if (!isBackgroundBlurSupported) {
-      throw new Error(
-        'Background blur is not supported. The processor may not be initialized yet.'
-      );
-    }
+    const currentProcessor = await initializeBackgroundBlur();
     try {
       const logger = options?.logger
-        ? options.logger
-        : new ConsoleLogger('BackgroundBlurProvider', LogLevel.INFO);
-      if (processor) {
+        || new ConsoleLogger('BackgroundBlurProvider', LogLevel.INFO);
+      if (currentProcessor) {
         const chosenVideoTransformDevice = new DefaultVideoTransformDevice(
           logger,
           selectedDevice,
-          [processor]
+          [currentProcessor]
         );
-        console.log(
-          `Created video transform device ${JSON.stringify(
-            chosenVideoTransformDevice,
-            null,
-            2
-          )}`
-        );
+        console.log(`Created video transform device ${JSON.stringify(chosenVideoTransformDevice, null, 2)}`);
         return chosenVideoTransformDevice;
       } else {
-        throw new Error('Processor has not been created.');
+        throw new Error('Processor has not been created. Background Blur is not supported.');
       }
     } catch (error) {
       throw new Error(