+++ src/utils/device-utils.ts
@@ -52,3 +52,15 @@ export const isOptionActive = (
 export const supportsSetSinkId = (): boolean => {
   return 'setSinkId' in HTMLAudioElement.prototype;
 };
+
+export function isPrevNextUndefined<T>(prev: T, next: T): boolean {
+  const isPrevUndefined = prev === undefined;
+  const isNextUndefined = next === undefined;
+  return isPrevUndefined && isNextUndefined;
+};
+
+export function isPrevNextEmpty<T>(prev: T, next: T): boolean {
+  const isPrevEmpty = (prev && Object.keys(prev).length === 0);
+  const isNextEmpty = (next && Object.keys(next).length === 0);
+  return isPrevEmpty && isNextEmpty;
+}